refactor(core): 用 BigDecimal 替换部分浮点数精度敏感变量和计算
- 将 AdaptiveBatchConfig 中 cpuThreshold、memThreshold、poolActiveThreshold 类型由 double 改为 BigDecimal，避免精度误差
- Agent 相关领域对象及服务中涉及模型温度参数由 double 改为 BigDecimal，实现统一高精度数值类型
- CostAnalysisService 优化为使用 BigDecimal 做价格计算，精度达到微美元级别
- CronjobMetrics 中系统负载相关权重和计算指标改为 BigDecimal，高精度负载分数计算
- DagParser 初始化集合容量优化，提升数据结构初始化效率
- DefaultAuditQueryService 增加防 SQL 注入严格校验和安全构造查询语句
- 多处日志静默异常捕获改为记录 DEBUG 级别日志，保障故障排查
- FeishuChannel appendSign 异常处理改为返回空 Map，避免空指针风险
- 删除 GeneratorThreadPoolConfig 配置类改为通过 ydsz.thread.pools.code-gen 配置统一管理线程池
- 代码生成服务注入线程池改为使用@Resource注解简化配置
- FileDedupService 中去重哈希键默认 TTL 改为静态常量，优化可维护性
- 统一线程池相关配置参数增加 ydsz.thread.pools.code-gen 线程池配置示例到 application.yml
- 其它小幅清理和注释完善，提升代码可读性和健壮性

+++ ydsz-agent/ydsz-agent-domain/src/main/java/com/njydsz/agent/domain/agent/AgentDefinition.java
@@ -1,6 +1,7 @@
 package com.njydsz.agent.domain.agent;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 
@@ -51,7 +52,7 @@ public final class AgentDefinition implements Serializable {
   private final Type type;
   private final String systemPrompt;
   private final List<String> toolNames;
-  private final double temperature;
+  private final BigDecimal temperature;
   private final int maxTokens;
   private final int maxIterations;
   private final String modelId;
@@ -63,7 +64,7 @@ public final class AgentDefinition implements Serializable {
       Type type,
       String systemPrompt,
       List<String> toolNames,
-      double temperature,
+      BigDecimal temperature,
       int maxTokens,
       int maxIterations,
       String modelId) {
@@ -105,7 +106,7 @@ public final class AgentDefinition implements Serializable {
     return toolNames;
   }
 
-  public double getTemperature() {
+  public BigDecimal getTemperature() {
     return temperature;
   }
